+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New `prepare:profile` command
 - New `prepare:test` command
 - New `prepare:mock-map-test` command
+- New `prepare:provider` command
 
 ## [2.0.0] - 2022-08-15
 ### Changed